Move alternation tests into their own suite.
