+++ MadnessTests/AlternationTests.swift
@@ -1,0 +1,26 @@
+//  Copyright (c) 2015 Rob Rix. All rights reserved.
+
+final class AlternationTests: XCTestCase {
+	let alternation = %"x" | (%"y" --> const(1))
+
+	func testAlternationParsesEitherAlternative() {
+		assertEqual(alternation("xy")?.1, "y")
+		assertEqual(alternation("yx")?.1, "x")
+	}
+
+	func testAlternationProducesTheParsedAlternative() {
+		assertEqual(alternation("xy")?.0, Either.left("x"))
+	}
+
+	func testAlternationOfASingleTypeCoalescesTheParsedValue() {
+		assertEqual((%"x" | %"y")("xy")?.0, "x")
+	}
+}
+
+
+// MARK: - Imports
+
+import Either
+import Madness
+import Prelude
+import XCTest
